requirements.txt
accelerate==0.30.0
antlr4-python3-runtime==4.9.3
bleach==6.1.0
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
diffusers==0.27.2
filelock==3.14.0
fsspec==2024.3.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.5.1
httpcore==0.9.1
httpx==0.23.0
huggingface-hub==0.23.0
hyperframe==5.2.0
idna==2.10
importlib-metadata==7.1.0
Jinja2==3.1.4
kaggle==1.6.12
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==24.0
pandas==2.0.3
pillow==10.3.0
protobuf==5.26.1
psutil==5.9.8
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
regex==2024.4.28
requests==2.32.0
rfc3986==1.5.0
safetensors==0.4.3
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
sympy==1.12
text-unidecode==1.3
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.40.1
triton==2.3.0
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
zipp==3.18.1
